Move files into the modeling folder.
