+++ src/execution/ast/type.cpp
@@ -3,7 +3,6 @@
 #include <unordered_map>
 #include <utility>
 
-#include "self_driving/operating_unit.h"
 #include "execution/exec/execution_context.h"
 #include "execution/sql/aggregation_hash_table.h"
 #include "execution/sql/aggregators.h"
@@ -17,6 +16,7 @@
 #include "execution/sql/thread_state_container.h"
 #include "execution/sql/value.h"
 #include "execution/sql/vector_projection_iterator.h"
+#include "self_driving/modeling/operating_unit.h"
 // #include "execution/util/csv_reader.h" Fix later.
 
 namespace noisepage::execution::ast {
