+++ src/execution/compiler/operator/operator_translator.cpp
@@ -1,6 +1,5 @@
 #include "execution/compiler/operator/operator_translator.h"
 
-#include "self_driving/operating_unit_util.h"
 #include "common/error/error_code.h"
 #include "common/error/exception.h"
 #include "common/json.h"
@@ -8,6 +7,7 @@
 #include "execution/compiler/work_context.h"
 #include "planner/plannodes/abstract_plan_node.h"
 #include "planner/plannodes/output_schema.h"
+#include "self_driving/modeling/operating_unit_util.h"
 #include "spdlog/fmt/fmt.h"
 
 namespace noisepage::execution::compiler {
