+++ src/include/metrics/pipeline_metric.h
@@ -13,8 +13,8 @@
 #include "common/resource_tracker.h"
 #include "metrics/abstract_metric.h"
 #include "metrics/metrics_util.h"
-#include "self_driving/operating_unit_util.h"
-#include "self_driving/operating_unit.h"
+#include "self_driving/modeling/operating_unit.h"
+#include "self_driving/modeling/operating_unit_util.h"
 #include "transaction/transaction_defs.h"
 
 namespace noisepage::metrics {
