+++ src/self_driving/modeling/operating_unit.cpp
@@ -1,4 +1,4 @@
-#include "self_driving/operating_unit.h"
+#include "self_driving/modeling/operating_unit.h"
 
 namespace noisepage::selfdriving {
